# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Simethris API Documentation",
        "description": "API untuk mendapatkan informasi SKL",
        "contact": {
            "email": "sultanjawer.work@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://siapriphonly.test/api",
            "description": "Simethris API Server"
        }
    ],
    "paths": {
        "/getToken": {
            "post": {
                "tags": [
                    "Get Access Token"
                ],
                "summary": "Post your username and password to get token. Use the token in the 'Authorization' header like so 'Bearer YOUR_TOKEN'",
                "description": "get token for Access API",
                "operationId": "getToken",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "user"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "mypass"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "The provided credentials are incorrect."
                    }
                }
            }
        },
        "/getSKL/{npwp}": {
            "get": {
                "tags": [
                    "SKL"
                ],
                "summary": "Get list of completed skl",
                "description": "Returns list of skl",
                "operationId": "getSKL",
                "parameters": [
                    {
                        "name": "npwp",
                        "in": "path",
                        "description": "npwp importir",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "simethrisToken": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Simethris",
            "description": "API Endpoints"
        }
    ],
    "components": {
        "securitySchemes": {
            "simethrisToken": {
                "type": "http",
                "scheme": "bearer",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}